refactor: extract hub link tests [ORB-10269]
